backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
marshmallow==3.20.1
psycopg2-binary==2.9.7
python-dotenv==1.0.0
gunicorn==21.2.0
pytest==7.4.2
requests==2.31.0
openai>=1.0.0
